# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAnimalNet API
  description: |
    Public JSON API for the OpenAnimalNet global animal data platform.

    All endpoints are rate-limited to **60 requests per minute per IP** and
    return `Cache-Control: public, s-maxage=60` headers.

    No API keys required — all data is derived from the 42-species dataset
    and live public APIs (GBIF, Wikidata, Wikipedia, iNaturalist).
  version: 1.23.0
  contact:
    name: OpenAnimalNet
    url: https://github.com/imredavid64-glitch/openanimalnet
  license:
    name: MIT

servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://openanimalnet.vercel.app
    description: Production

paths:
  /api/v1/animals:
    get:
      operationId: listAnimals
      summary: List species
      description: |
        List all species with optional filtering and pagination.
        Returns a paginated array of animal profiles.
      tags: [Animals]
      parameters:
        - name: category
          in: query
          schema:
            type: string
            enum: [mammals, birds, reptiles, amphibians, fish, invertebrates, insects, marine]
          description: Filter by animal category
        - name: conservationStatus
          in: query
          schema:
            type: string
            enum: [EX, EW, CR, EN, VU, NT, LC, DD, NE]
          description: Filter by IUCN conservation status
        - name: dataCategories
          in: query
          schema:
            type: string
          description: Comma-separated data category filter (biological, behavioral, ecological, population, health, agricultural, shelter, human-interaction)
          example: biological,health
        - name: isMonitored
          in: query
          schema:
            type: boolean
          description: Filter by monitoring status
        - name: search
          in: query
          schema:
            type: string
          description: Full-text search across common name, scientific name, and description
          example: elephant
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
          description: Page number
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: Results per page
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Animal"
                      pagination:
                        $ref: "#/components/schemas/Pagination"
        "429":
          $ref: "#/components/responses/RateLimited"

  /api/v1/animals/{id}:
    get:
      operationId: getAnimal
      summary: Get species profile
      description: Full profile for one species, including all five data categories.
      tags: [Animals]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Species ID (e.g., elephant-001, lion-001)
          example: lion-001
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AnimalData"
        "404":
          description: Species not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/RateLimited"

  /api/v1/populations:
    get:
      operationId: listPopulations
      summary: Population estimates
      description: Population estimates and conservation metrics for all species.
      tags: [Populations]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/PopulationRecord"
        "429":
          $ref: "#/components/responses/RateLimited"

  /api/v1/monitoring/alerts:
    get:
      operationId: listAlerts
      summary: List monitoring alerts
      description: Active monitoring alerts, filterable by severity type.
      tags: [Monitoring]
      parameters:
        - name: type
          in: query
          schema:
            type: string
            enum: [critical, warning, info]
          description: Filter by alert severity
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/Alert"
        "400":
          description: Invalid type parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/RateLimited"

  /api/v1/monitoring/stats:
    get:
      operationId: getMonitoringStats
      summary: Dashboard statistics
      description: Aggregated monitoring statistics for the dashboard.
      tags: [Monitoring]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/MonitoringStats"
        "429":
          $ref: "#/components/responses/RateLimited"

  /api/v1/locations:
    get:
      operationId: listLocations
      summary: Recent telemetry locations
      description: Recent tracking locations for all monitored animals.
      tags: [Locations]
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/LocationRecord"
        "429":
          $ref: "#/components/responses/RateLimited"

  /api/v1/live/sync:
    get:
      operationId: liveSync
      summary: Live GBIF sync
      description: |
        Fetch recent georeferenced occurrences from GBIF for a specific species.
        Results are cached for 60 seconds per species.
      tags: [Live Sync]
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
          description: Species ID
          example: lion-001
      responses:
        "200":
          description: Success
          headers:
            Cache-Control:
              schema:
                type: string
                example: public, max-age=60, stale-while-revalidate=300
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/GbifSync"
        "400":
          description: Missing id parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: GBIF sync failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/RateLimited"

  /api/v1/live/observations:
    get:
      operationId: liveObservations
      summary: Batched live GBIF observations
      description: |
        Fetch recent georeferenced occurrences from GBIF for a batch of
        species (the globe's live-observations layer). Each species is
        cached for 60 seconds; a species that fails or has no GBIF key is
        skipped rather than failing the whole request.
      tags: [Live Sync]
      parameters:
        - name: ids
          in: query
          required: true
          schema:
            type: string
          description: Comma-separated species IDs
          example: lion-001,tiger-001,blue-whale-001
      responses:
        "200":
          description: Success
          headers:
            Cache-Control:
              schema:
                type: string
                example: public, max-age=60, stale-while-revalidate=300
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/LiveObservations"
        "400":
          description: Missing or too many ids
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "502":
          description: GBIF observations failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/RateLimited"

  /api/v1/live/observations:
    post:
      operationId: submitObservation
      summary: Submit a crowd-sourced observation
      description: |
        Accept a wildlife observation (replayed by the service worker's
        background sync when connectivity returns). Broadcasts an SSE alert.
      tags: [Live Sync]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [animalId]
              properties:
                animalId: { type: string }
                species: { type: string }
                latitude: { type: number }
                longitude: { type: number }
                notes: { type: string }
                reportedBy: { type: string }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessResponse"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/SubmittedObservation"
        "400":
          description: Invalid fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          $ref: "#/components/responses/RateLimited"

  /api/v1/search:
    get:
      operationId: universalSearch
      summary: Universal species search
      description: Search across GBIF, Wikipedia, Wikidata, and iNaturalist.
      tags: [Search]
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query
      responses:
        "200":
          description: Success
        "400":
          description: Missing q parameter
        "502":
          description: Upstream search failed
        "429":
          $ref: "#/components/responses/RateLimited"

  /api/v1/identify:
    post:
      operationId: identifySpecies
      summary: Identify a species from an image
      tags: [Identify]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [image]
              properties:
                image: { type: string, description: Base64 or data-URL image }
                topK: { type: integer, description: Max results, default: 5 }
      responses:
        "200":
          description: Matches returned
        "400":
          description: Missing image field
        "429":
          $ref: "#/components/responses/RateLimited"

  /api/v1/subscriptions:
    get:
      operationId: listSubscriptions
      summary: List alert subscriptions
      tags: [Subscriptions]
      parameters:
        - name: email
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Success
        "429":
          $ref: "#/components/responses/RateLimited"
    post:
      operationId: createSubscription
      summary: Create an alert subscription
      tags: [Subscriptions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string }
                speciesIds: { type: array, items: { type: string } }
                regions:
                  type: array
                  items:
                    type: object
                    properties:
                      lat: { type: number }
                      lng: { type: number }
                      radiusKm: { type: number }
                alertTypes:
                  type: array
                  items: { type: string, enum: [critical, warning, info] }
      responses:
        "201":
          description: Created
        "400":
          description: Invalid email
        "429":
          $ref: "#/components/responses/RateLimited"
    delete:
      operationId: deleteSubscription
      summary: Remove a subscription
      tags: [Subscriptions]
      parameters:
        - name: id
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
        "400":
          description: Missing id
        "404":
          description: Not found
        "429":
          $ref: "#/components/responses/RateLimited"

  /api/v1/annotations:
    get:
      operationId: listAnnotations
      summary: List annotations
      tags: [Annotations]
      parameters:
        - name: animalId
          in: query
          schema: { type: string }
      responses:
        "200":
          description: Success
        "429":
          $ref: "#/components/responses/RateLimited"
    post:
      operationId: createAnnotation
      summary: Create an annotation
      tags: [Annotations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [animalId, author, content]
              properties:
                animalId: { type: string }
                author: { type: string }
                content: { type: string }
                category: { type: string, enum: [observation, correction, note, sighting] }
                location:
                  type: object
                  properties:
                    lat: { type: number }
                    lng: { type: number }
      responses:
        "201":
          description: Created
        "400":
          description: Invalid fields
        "429":
          $ref: "#/components/responses/RateLimited"
    delete:
      operationId: deleteAnnotation
      summary: Remove an annotation
      tags: [Annotations]
      parameters:
        - name: id
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
        "404":
          description: Not found
        "429":
          $ref: "#/components/responses/RateLimited"

  /api/v1/webhooks:
    get:
      operationId: listWebhooks
      summary: List registered webhooks
      tags: [Webhooks]
      responses:
        "200":
          description: Success
        "429":
          $ref: "#/components/responses/RateLimited"
    post:
      operationId: registerWebhook
      summary: Register a webhook
      tags: [Webhooks]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url: { type: string }
                events:
                  type: array
                  items: { type: string }
      responses:
        "201":
          description: Created
        "400":
          description: Invalid URL
        "429":
          $ref: "#/components/responses/RateLimited"
    delete:
      operationId: deleteWebhook
      summary: Remove a webhook
      tags: [Webhooks]
      parameters:
        - name: id
          in: query
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deleted
        "404":
          description: Not found
        "429":
          $ref: "#/components/responses/RateLimited"

  /api/v1/export:
    get:
      operationId: exportSpecies
      summary: Export the species dataset
      tags: [Export]
      parameters:
        - name: format
          in: query
          schema: { type: string, enum: [geojson, csv, kml], default: geojson }
        - name: category
          in: query
          schema: { type: string }
      responses:
        "200":
          description: File download
          content:
            application/json: {}
            text/csv: {}
            application/vnd.google-earth.kml+xml: {}
        "429":
          $ref: "#/components/responses/RateLimited"

  /api/v1/feed:
    get:
      operationId: alertFeed
      summary: Conservation-alert feed
      tags: [Feed]
      parameters:
        - name: format
          in: query
          schema: { type: string, enum: [atom, rss], default: atom }
      responses:
        "200":
          description: XML feed
          content:
            application/atom+xml: {}
            application/rss+xml: {}

  /api/v1/live/alerts:
    get:
      operationId: liveAlertsStream
      summary: Live alert SSE stream
      description: Server-sent-events stream of live conservation alerts.
      tags: [Live Sync]
      responses:
        "200":
          description: text/event-stream

components:
  schemas:
    Animal:
      type: object
      properties:
        id:
          type: string
          example: lion-001
        commonName:
          type: string
          example: African Lion
        scientificName:
          type: string
          example: Panthera leo
        category:
          type: string
          enum: [mammals, birds, reptiles, amphibians, fish, invertebrates, insects, marine]
        description:
          type: string
        images:
          type: array
          items:
            type: string
        conservationStatus:
          type: string
          enum: [EX, EW, CR, EN, VU, NT, LC, DD, NE]
        taxonomy:
          type: object
          properties:
            kingdom: { type: string }
            phylum: { type: string }
            class: { type: string }
            order: { type: string }
            family: { type: string }
            genus: { type: string }
            species: { type: string }
        location:
          $ref: "#/components/schemas/GeoPoint"
        habitat:
          type: array
          items:
            type: string
        populationEstimate:
          type: integer
        populationHistory:
          type: array
          items:
            $ref: "#/components/schemas/PopulationPoint"
        populationHistoryNote:
          type: string
        isMonitored:
          type: boolean
        lastUpdated:
          type: string
          format: date-time
        dataCategories:
          type: array
          items:
            type: string
        migrationRoutes:
          type: array
          items:
            $ref: "#/components/schemas/MigrationRoute"

    AnimalData:
      type: object
      properties:
        animal:
          $ref: "#/components/schemas/Animal"
        biological:
          type: object
        behavioral:
          type: object
        ecological:
          type: object
        population:
          type: object
        health:
          type: object

    GeoPoint:
      type: object
      properties:
        latitude: { type: number }
        longitude: { type: number }
        altitude: { type: number }
        accuracy: { type: number }
        timestamp: { type: string, format: date-time }
        source: { type: string }

    PopulationPoint:
      type: object
      properties:
        year: { type: integer }
        estimate: { type: integer }

    MigrationRoute:
      type: object
      properties:
        name: { type: string }
        season: { type: string }
        startMonth: { type: integer }
        endMonth: { type: integer }
        waypoints:
          type: array
          items:
            $ref: "#/components/schemas/GeoPoint"

    PopulationRecord:
      type: object
      properties:
        animalId: { type: string }
        commonName: { type: string }
        scientificName: { type: string }
        conservationStatus: { type: string }
        populationEstimate: { type: integer, nullable: true }
        aerialSurveyCounts: { type: number, nullable: true }
        cameraTrapCaptureRates: { type: number, nullable: true }
        rangeContractionPercentage: { type: number, nullable: true }

    Alert:
      type: object
      properties:
        id: { type: string }
        animalId: { type: string }
        type:
          type: string
          enum: [critical, warning, info]
        title: { type: string }
        message: { type: string }
        location:
          $ref: "#/components/schemas/GeoPoint"
        timestamp: { type: string, format: date-time }

    MonitoringStats:
      type: object
      properties:
        totalAnimals: { type: integer }
        monitoredAnimals: { type: integer }
        activeAlerts: { type: integer }
        monitoringCoverage:
          type: object
          properties:
            mammals: { type: number }
            birds: { type: number }
            reptiles: { type: number }
            amphibians: { type: number }
            fish: { type: number }
            marine: { type: number }

    LocationRecord:
      type: object
      properties:
        animalId: { type: string }
        commonName: { type: string }
        latitude: { type: number }
        longitude: { type: number }
        altitude: { type: number, nullable: true }
        accuracy: { type: number, nullable: true }
        source: { type: string }
        timestamp: { type: string, format: date-time }

    GbifSync:
      type: object
      properties:
        animalId: { type: string }
        scientificName: { type: string }
        gbifKey: { type: integer }
        occurrences:
          type: array
          items:
            type: object
            properties:
              key: { type: integer }
              decimalLatitude: { type: number }
              decimalLongitude: { type: number }
              country: { type: string }
              eventDate: { type: string }
              basisOfRecord: { type: string }
        fetchedAt: { type: string, format: date-time }
        count: { type: integer }

    LiveObservations:
      type: object
      properties:
        fetchedAt: { type: string, format: date-time }
        cached: { type: boolean }
        skipped:
          type: array
          items: { type: string }
        observations:
          type: array
          items:
            type: object
            properties:
              animalId: { type: string }
              key: { type: integer }
              species: { type: string }
              country: { type: string, nullable: true }
              latitude: { type: number }
              longitude: { type: number }
              eventDate: { type: string, nullable: true }

    SubmittedObservation:
      type: object
      properties:
        id: { type: string }
        animalId: { type: string }
        species: { type: string, nullable: true }
        latitude: { type: number, nullable: true }
        longitude: { type: number, nullable: true }
        notes: { type: string, nullable: true }
        reportedBy: { type: string, nullable: true }
        timestamp: { type: string, format: date-time }

    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true

    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        message:
          type: string

    Pagination:
      type: object
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer }

  responses:
    RateLimited:
      description: Rate limit exceeded (60 req/min per IP)
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
      content:
        application/json:
          schema:
            allOf:
              - $ref: "#/components/schemas/ErrorResponse"
              - type: object
                properties:
                  retryAfter:
                    type: integer
          example:
            success: false
            error: "Rate limit exceeded"
            retryAfter: 45

  tags:
    - name: Animals
      description: Species profiles and listings
    - name: Populations
      description: Population estimates and conservation metrics
    - name: Monitoring
      description: Alerts and dashboard statistics
    - name: Locations
      description: Telemetry tracking data
    - name: Live Sync
      description: Real-time GBIF occurrence data
    - name: Search
      description: Universal species search
    - name: Identify
      description: Image-based species identification
    - name: Subscriptions
      description: Alert subscriptions
    - name: Annotations
      description: Per-species notes and sightings
    - name: Webhooks
      description: Outbound alert webhooks
    - name: Export
      description: Full dataset export
    - name: Feed
      description: Atom/RSS conservation feeds